So first up, sorry for spamming PRs :D

As mentioned in #399 , parsing a very large stylesheet (taken from https://materialdesignicons.com/ ) made the application startup noticeably slower. 

By turning the StyleSheetNode children list from a vector into a sorted set using the comparison methods for the CompoundSelector members, i was able to speedup the application startup (time to first render) from 2.3s to 1.0s.