@mikke89 2022-12-30
Sounds pretty good!

Is it possible to use a hash map instead of a set?

I'd like to see more benchmarks. Can you replicate this situation in a benchmark? I'd also like to see the effect of this on more trivial cases (smaller style sheets).

I did some quick benchmarking, and there seems to be a measurable performance pessimization for the ElementDocument benchmark, as seen below.

#### This PR

| relative |               us/op |                op/s |    err% |     total | ElementDocument
|---------:|--------------------:|--------------------:|--------:|----------:|:----------------
|   100.0% |              490.50 |            2,038.74 |    2.3% |      0.01 | `LoadDocument`
|    98.1% |              500.00 |            2,000.00 |    2.2% |      0.01 | `LoadDocument + Show`
|    98.1% |              500.20 |            1,999.20 |    0.9% |      0.01 | `LoadDocument + Show + Update`
|    87.7% |              559.30 |            1,787.95 |    3.6% |      0.01 | `LoadDocument + Show + Update + Render`

| relative |               us/op |                op/s |    err% |     total | ElementDocument w/ClearStyleSheetCache
|---------:|--------------------:|--------------------:|--------:|----------:|:---------------------------------------
|   100.0% |            1,502.90 |              665.38 |    2.5% |      0.02 | `Clear + LoadDocument`
|   101.5% |            1,481.40 |              675.04 |    2.3% |      0.02 | `Clear + LoadDocument + Show`
|    99.4% |            1,511.60 |              661.55 |    2.3% |      0.02 | `Clear + LoadDocument + Show + Update`
|    98.7% |            1,522.00 |              657.03 |    1.7% |      0.02 | `Clear + LoadDocument + Show + Update + Render`


#### Master

| relative |               us/op |                op/s |    err% |     total | ElementDocument
|---------:|--------------------:|--------------------:|--------:|----------:|:----------------
|   100.0% |              440.90 |            2,268.09 |    3.9% |      0.00 | `LoadDocument`
|   100.7% |              438.00 |            2,283.11 |    0.8% |      0.00 | `LoadDocument + Show`
|    94.4% |              467.30 |            2,139.95 |    2.3% |      0.01 | `LoadDocument + Show + Update`
|    84.8% |              520.10 |            1,922.71 |    2.7% |      0.01 | `LoadDocument + Show + Update + Render`

| relative |               us/op |                op/s |    err% |     total | ElementDocument w/ClearStyleSheetCache
|---------:|--------------------:|--------------------:|--------:|----------:|:---------------------------------------
|   100.0% |            1,393.10 |              717.82 |    2.5% |      0.02 | `Clear + LoadDocument`
|   101.5% |            1,373.10 |              728.28 |    1.6% |      0.02 | `Clear + LoadDocument + Show`
|   101.6% |            1,371.00 |              729.39 |    2.0% |      0.02 | `Clear + LoadDocument + Show + Update`
|    96.7% |            1,440.20 |              694.35 |    0.9% |      0.02 | `Clear + LoadDocument + Show + Update + Render`