@Dakror 2022-12-30
Currently the system does a == operator for all objects. With my approach, it only does it log(n) times due to using std::set. Hashing and thus reducing the lookup time even further is very difficult since the class names and all that stuff are user-generated. You'd probably need several dictionaries or whatever to compute a truly stable and unique-enough hash. For icon fonts with about 20k lines of CSS, such as the aforementioned material design icons, this slowdown is quite significant.
  
But the slowdown you're observing probably comes from the truly bad implementation of having to allocate a unique ptr just for the find call and potentially discarding it. I would either like to setup some better finding method, or best case find a way to consistently hash the compound selector, yeah