@Dakror 2022-12-30
I'll look into the performance degradation. 

I would also much rather use a hash map but the issue lies in finding a consistent hashing function for the CompoundSelector struct.