@mikke89 2022-12-30
Yeah, it's not a given that hashing might not be faster always. We do have a `HashCombine`  utility which might be helpful, if you want to try out this route.

I'm kind of wishing that users don't add 4000-lines of CSS font code. It's possible that we might find some worthwhile optimization opportunities for common cases when testing this, which I'd be fully onboard with. But if it comes to pessimization in common cases, or adding complexity for handling extreme cases, I'd rather keep it basic and expect some more manual authoring from users.