@Dakror 2022-12-30
Yeah i completely understand that sentiment.