@mikke89 2022-12-30
Right, I don't think an icon font necessarily is a big issue. But you probably won't use all of those 4000 icons, so I would rather manually add the CSS rules for the ones being used. Or, since we don't support the `content` property right now, just add the character to the RML. In many cases I think this is better anyway, especially for one-off icons. It would be nice to support html hex character encoding for this to be a bit more ergonomic.