@Dakror 2022-12-31
okay i can totally live with that. So then let's close this and the other PR for now, and I'll instead add html-based escaping